feat: separate each view state and add upload file from gallery
